[FLG] Make max distance query param double

+++ src/main/java/ca/ulaval/glo2003/beds/domain/BedQueryBuilder.java
@@ -27,7 +27,7 @@ public interface BedQueryBuilder {
 
   BedQueryBuilder withOrigin(Location origin);
 
-  BedQueryBuilder withMaxDistance(int maxDistance);
+  BedQueryBuilder withMaxDistance(double maxDistance);
 
   InMemoryBedQuery build();
 }
